Moved AxiosError message formatting to own function
