+++ dataland-frontend/src/components/forms/CreateEUTaxonomyForFinancials.vue
@@ -323,7 +323,7 @@ import {
   EuTaxonomyDataForFinancialsFinancialServicesTypesEnum,
   EuTaxonomyDataForNonFinancials,
 } from "@clients/backend";
-import { AxiosError, AxiosResponse } from "axios";
+import { AxiosResponse } from "axios";
 import { modifyObjectKeys, ObjectType, updateObject } from "@/utils/updateObjectUtils";
 import { formatBytesUserFriendly } from "@/utils/NumberConversionUtils";
 import JumpLinksSection from "@/components/forms/parts/JumpLinksSection.vue";
@@ -331,6 +331,7 @@ import DataPointForm from "@/components/forms/parts/kpiSelection/DataPointForm.v
 import SubmitButton from "@/components/forms/parts/SubmitButton.vue";
 import { FormKitNode } from "@formkit/core";
 import UploadReports from "@/components/forms/parts/UploadReports.vue";
+import { formatAxiosErrorMessage } from "@/utils/AxiosErrorMessageFormatter";
 
 export default defineComponent({
   setup() {
@@ -544,21 +545,7 @@ export default defineComponent({
       } catch (error) {
         this.messageCount++;
         console.error(error);
-
-        if (error instanceof AxiosError) {
-          const err = error as AxiosError;
-          const response = err.response as AxiosResponse<{ errors: { summary: string; message: string }[] }>;
-          let errSummary = JSON.stringify(err);
-          let errMessage = "";
-
-          if (response.data.errors[0]) {
-            errSummary = response.data.errors[0].summary;
-            errMessage = response.data.errors[0].message;
-          }
-          this.message = `${errSummary} ${errMessage}`;
-        } else {
-          this.message = (error as Error).message;
-        }
+        this.message = formatAxiosErrorMessage(error);
       } finally {
         this.postEuTaxonomyDataForFinancialsProcessed = true;
       }
